+++ popup.html
@@ -11,7 +11,8 @@ body{height: 190px; width: 150px; margin: 1%; padding: 10px;}
     <script src="background.js"></script>
     <script src="options.js"></script>
     <script src="bootstrap.min.js"></script>
-    <script src="bootbox.all.js"></script>
+    <script src="bootbox.js"></script>
+    <script src="jquery-3.4.1.min.js"></script>
     <link rel="stylesheet" href="bootstrap.min.css">
 </head>
 <body>
